move custom mapping to pageobjectclient away from componentfactory
